Add a new param: num_files_to_store in dropbox-sdk
Add the param num_files_to_store to the config.yml file in order to save just the last backups in the dropbox folder, and free space deleting the old copies. 

That is for big projects that used too much disk. If you wanna store the the last week for example, using a diary backup, you should use this configuration:

        dropbox_sdk:
            remote_path: %dropbox_path% 
            access_token: %dropbox_token% 
            num_files_to_store: 7 # Number of files you wanna store in the backup folder

The default configuration is 0, and that means that the script won't delete any file.